chore: migrate commands
