--- cmd/demotest/internal/root.go
@@ -1,29 +1,0 @@
-package internal
-
-import (
-	"os"
-
-	"github.com/spf13/cobra"
-)
-
-// rootCmd represents the base command when called without any subcommands
-var rootCmd = &cobra.Command{
-	Use:   "demotest",
-	Short: "A tool that runs all demo",
-	Long: `This application is a tool that integrates llpkg-related functionality,
-such as binary installation, configuration file parsing, etc.
-`,
-}
-
-// Execute adds all child commands to the root command and sets flags appropriately.
-// This is called by main.main(). It only needs to happen once to the rootCmd.
-func Execute() {
-	err := rootCmd.Execute()
-	if err != nil {
-		os.Exit(1)
-	}
-}
-
-func init() {
-	rootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-}
