--- cmd/demotest/main.go
@@ -1,7 +1,0 @@
-package main
-
-import cmd "github.com/goplus/llpkgstore/cmd/demotest/internal"
-
-func main() {
-	cmd.Execute()
-}
