--- cmd/maintain/main.go
@@ -1,7 +1,0 @@
-package main
-
-import cmd "github.com/goplus/llpkgstore/cmd/maintain/internal"
-
-func main() {
-	cmd.Execute()
-}
